Player fixes ; replace radix AspectRatio with mantine
