+++ package.json
@@ -19,12 +19,11 @@
   },
   "dependencies": {
     "@eluvio/elv-client-js": "^4.0.92",
-    "@eluvio/elv-player-js": "^2.0.11",
+    "@eluvio/elv-player-js": "^2.0.19",
     "@mantine/core": "7.6.1",
     "@mantine/hooks": "7.6.1",
     "@mantine/notifications": "7.6.1",
     "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-aspect-ratio": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.6",
     "@tabler/icons-react": "^2.34.0",
