--- src/components/AspectRatio.jsx
@@ -1,11 +1,0 @@
-import * as AspectRatioPrimitive from "@radix-ui/react-aspect-ratio";
-
-const AspectRatio = ({ratio=16/9, children}) => {
-  return (
-    <AspectRatioPrimitive.Root ratio={ratio}>
-      {children}
-    </AspectRatioPrimitive.Root>
-  );
-};
-
-export default AspectRatio;
