+++ src/pages/monitor/Monitor.jsx
@@ -1,97 +1,13 @@
-import {useEffect, useState} from "react";
+import {useState} from "react";
 import {observer} from "mobx-react-lite";
-import {IconX} from "@tabler/icons-react";
-import {ActionIcon, TextInput} from "@mantine/core";
+import {TextInput} from "@mantine/core";
 import {useDebouncedValue} from "@mantine/hooks";
 
 import {dataStore, streamStore} from "@/stores";
-import AspectRatio from "@/components/AspectRatio.jsx";
-import Video from "@/components/Video.jsx";
 import {Loader} from "@/components/Loader.jsx";
 import {SortTable} from "@/utils/helpers";
-import {PlayCircleIcon as PlayIcon} from "@/assets/icons";
 import {STATUS_TEXT} from "@/utils/constants";
-
-export const VideoContainer = observer(({slug, index, showPreview}) => {
-  const [play, setPlay] = useState(false);
-  const [frameKey, setFrameKey] = useState(0);
-  const [frameSegmentUrl, setFrameSegmentUrl] = useState(streamStore.streamFrameUrls[slug]?.url);
-  const status = streamStore.streams?.[slug]?.status;
-
-  useEffect(() => {
-    if(!showPreview || play || status !== "running") {
-      return;
-    }
-
-    const existingFrame = streamStore.streamFrameUrls[slug];
-    // Frame loading already initialized - no delay needed
-    if(frameKey > 0 || (existingFrame && Date.now() - existingFrame.timestamp < 60000)) {
-      setFrameSegmentUrl(existingFrame.url);
-      // eslint-disable-next-line no-console
-      console.log("SKIP DELAY", slug);
-      return;
-    }
-
-    // Stagger frame loads
-    const delay = Math.min(200 + 500 * index, 10000);
-    const frameTimeout = setTimeout(async () => {
-      setFrameSegmentUrl(await streamStore.StreamFrameURL(slug));
-    }, delay);
-
-    return () => clearTimeout(frameTimeout);
-  }, [play, frameKey, status, showPreview]);
-
-  // Reload frame every minute after initial frame load
-  useEffect(() => {
-    if(!frameSegmentUrl) { return; }
-
-    const delay = 60000 + Math.min(200 + 500 * index, 10000);
-
-    const updateTimeout = setTimeout(() => setFrameKey(frameKey + 1), delay);
-
-    return () => clearTimeout(updateTimeout);
-  }, [frameKey, frameSegmentUrl]);
-
-  return (
-    <div className="monitor__video-wrapper">
-      <AspectRatio>
-        {
-          !play ?
-            <button
-              role="button"
-              tabIndex={1}
-              onClick={() => setPlay(true)}
-              className="monitor__video-placeholder"
-            >
-              <PlayIcon width={45} height={45} />
-              {
-                (!showPreview || !frameSegmentUrl) ? null :
-                  <video src={frameSegmentUrl} className="monitor__video-frame" controls={false} onContextMenu={e => e.preventDefault()} />
-              }
-            </button> :
-            <>
-              <Video
-                objectId={streamStore.streams[slug].objectId}
-                playerOptions={{
-                  capLevelToPlayerSize: false,
-                  autoplay: true
-                }}
-              />
-              <ActionIcon
-                className="monitor__video-x"
-                title="Stop Playback"
-                color="gray.1"
-                variant="transparent"
-                onClick={() => setPlay(false)}
-              >
-                <IconX />
-              </ActionIcon>
-            </>
-        }
-      </AspectRatio>
-    </div>
-  );
-});
+import VideoContainer from "@/components/VideoContainer.jsx";
 
 const Monitor = observer(() => {
   const [filter, setFilter] = useState("");
